Added new verb Purchased
ADL suggested that this should be in this profile as opposed to the pending GBLxAPI profile.  We are currently using this verb in xAPI collection in games.  We believe this should be a community verb as it is a common activity.  We have followed your format which may or may not meet current requirements.